% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sbhm.R
\name{sbhm}
\alias{get.sbhm.summary}
\alias{plot.sbhm}
\alias{print.sbhm}
\alias{print.summary.sbhm}
\alias{sbhm}
\alias{summary.sbhm}
\title{Fit a simple hierarchical normal model with MCMC}
\usage{
sbhm(y, s, m = 0, tauScale = NULL, engine = c("stan", "jags"),
  nchains = NULL, iter = 41000, burn = 1000, thin = 4,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975))

\method{print}{sbhm}(x, digits = 3, ...)

\method{plot}{sbhm}(x, xlab = "Estimates", ylab = "", main = "",
  mle.col = "blue", sbhm.col = "orange", margins = c(5.1, 7.1, 4.1, 2.1),
  offset = 0.2, lwd.step = 2, cex = 1.5, ...)
}
\arguments{
\item{y}{The observed group means.}

\item{s}{The observed group standard errors.}

\item{m}{The top level mean. Defaults to \code{m=0}.}

\item{tauScale}{The scale parameter for the half-Cauchy distribution for the top
level scale. Defaults to twice the range of \code{y}.}

\item{engine}{Which MCMC engine to use. Defaults to \code{engine="stan"} and
if \code{stan} is not avaiable will try \code{JAGS}. Note that if the engine
is JAGS, a since Markov chain will be run, irrespective of the \code{nchains}
argument.}

\item{nchains}{The number of chains to run. If not specified and \code{stan} is
available, the function tries to figure out how many cores are available,
then uses all but one of them to run a chain. If JAGS is to be used, it defaults
to running 3 chains in sequence, not parallel.}

\item{iter}{The number of steps in each Markov chain. Defaults to \code{iter=41000}.}

\item{burn}{The number of burn-in steps to be discarded from each chain.
Defaults to \code{burn=1000}.}

\item{thin}{The amount of thinning of each chain to do. Defaults to \code{thin=4}.}

\item{probs}{The quantiles of the posterior distribution to print. Defaults to
\code{probs=c(.025, .25, .5, .75, .975)}.}

\item{x}{In \code{plot.sbhm}, an object of class "sbhm".}

\item{digits}{In \code{print.sbhm}, the number of digits to round to.}

\item{...}{Additional arguments to \code{plot}. Not used.}

\item{xlab}{In \code{plot.sbhm}, the x-axis label.}

\item{ylab}{In \code{plot.sbhm}, the y-axis label.}

\item{main}{In \code{plot.sbhm}, the main title.}

\item{mle.col}{In \code{plot.sbhm}, the colour of the mles.}

\item{sbhm.col}{In \code{plot.sbhm}, the colour of the shrunken estimates.}

\item{margins}{Vector of length 4 giving the margin sizes for \code{plot}. Defaults
to \code{margins=c(5.1, 7.1, 4.1, 2.1)}.}

\item{offset}{In \code{plot.sbhm}, how much to nudge the mle estiamtes upwards and the
sbhm estimates downwards. Defaults to \code{offset=.2}.}

\item{lwd.step}{The difference in line widths for the confidence intervals. Defaults
to \code{lwd.step=2}.}

\item{cex,}{In \code{plot.sbhm}, character expansion. Defaults to \code{cex=1.5}.}
}
\value{
An object of class "sbhm" containing the fitted model, \code{probs},
  \code{engine}, the data and function call.
}
\description{
Fit a simple hierarchical normal model with MCMC
}
\details{
The function requires that you have either rstan or rjags installed.
  Because many users of rstan would have no need for rjags, rjags is not
  featured as a dependency. Because rstan is not on CRAN and takes a little
  effort to install, the user might want to use rjags instead, so rstan is
  not featured as a dpeendency.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print method for sbhm

\item \code{plot}: Plot an sbhm object
}}
\examples{
\dontrun{
mod <- sbhm(schools$effect, schools$se)
plot(mod)
# The MERIT-HF study
y <- merit$logHR
names(y) <- merit$country # used as labels in the plot
mod <- sbhm(y=y, s=merit$se, nchains=2)
plot(mod)
}
}
\author{
Harry Southworth
}

